eighth commit

+++ c/function.h
@@ -1,4 +1,4 @@
 #include <stdio.h>
 
 void calcul(int a, int b);
-long factoriel(int n);
+long double factoriel(int n);
